prote/drf-prote-test/apps/app_crud/serializers/user_role_serializer.py
from rest_framework import serializers

from apps.app_crud.models import UserRole


class UserRoleSerializer(serializers.ModelSerializer):

    class Meta:
        model = UserRole
        fields = '__all__'
